[FCE-313] Add reconnection feature (#57)
This PR adds a feature for automatic reconnection. If there is no
internet the sdk will attempt to connect again with linear backoff. User
can configure the amount of attempts and the delay. User can disable
this by setting max attempts to 0.
When reconnecting we must "recreate" local tracks (if we don't do this,
we'll crash the backend with duplicated tracks error). But to avoid
opening up the camera and all this mess with screencast permissions etc.
we're reusing track sources, so the experience is smoother.
Also app developer can use a hook to keep track of reconnection status
and inform app users what's happening.