[FCE-402] Use branded type as camera id (#83)
## Description

Branded types are nice tool for restricting type of values that can be
passed into functions.
This change converts camera id into such type. This way, when switching
camera, developer won't be able to pass arbitrary string. It will be
possible to pass only id we are allowing to use.

## Motivation and Context

Add better restriction to API


## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)