[FCE-502] Fix Android crashing on sdkVersion >= 34 (#105)
## Description

The whole issue is related to permissions for `mediaProjection`. Somehow
the previous implementation (even though it looked fine) requested the
foreground service with `mediaProjection` on launch, without permission
first.

- Added an Expo Module for Android which handles starting a foreground
service for the user (required for Android >= 34)
- Added an Expo Config Plugin to allow conditional enabling of service

**Discussion**: I believe, ideally, to provide the best experience for
our users, we should start the service implicitly inside the SDK after
requesting screen sharing permissions (if the user has
`enableForegroundService` set to true in our plugin). Right now there is
a `if == android` in our example which is not ideal and not relevant for
most users.

## Motivation and Context

The previous implementation that used Notifee was removed with the Expo
upgrade.

## How has this been tested?

- Successfully launched Screen sharing on Android 34
- Launched iOS Room Screen with no issues

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.