[FCE-30] fix vad (#174)
## Description

Fix vad and utilize it in fishjam-chat. VAD status is taken from
fishjam, but i discovered it is possible to implement it on native side
by working with RTCAudioTrack. Currently when single VAD status changes
all endpoints are emitted. This solution might be expensive but it is
very simple and it works. In future it might be a good idea to just emit
single changed track.

## Motivation and Context

VAD is cool

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.