[FCE-621] Add jitpack cache warmup (#176)
## Description

When we publish new release (tag), we have to push libraries to
CocoaPods and npm. But jitpack works another way, and it pulls our code
from repo.

This PR is adding step to let know Jitpack that new version was
released.

This PR also adds formatting for .github files.

## Motivation and Context

Right now, when we publish new version, first user that will try to
build app will get error - as Android won't work. There are few minutes
required after first request for JitPack to cache everything. With this
change, it should work right away.

## How has this been tested?

To fully test this addition, new version has to be released. But I
should not break any existing flow.

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.

## Screenshots (if appropriate)