[FCE-555] Add ex-webrtc support (#177)
## Description

This PR adds support for ex-webrtc backend. 
On both platforms:
- ~removed `restartIce()` call - it causes issues with connectivity~ =>
`restartIce` is needed for non exwebrtc.
- changes ice transport type from `.relay` to .`all`
- conforms to protocol when sending candidates
- disables simulcast in example app - it's currently not supported, but
maybe it will be supported in the future, thus I'm not changing the api,
it's disabled by default anyway

NEW changes:
- Removed `endpointType` as it was unused and only created confusion
about what should be set.
- ExWebRTC/WebRTC is now decided based upon if turn servers are passed
(same as it's done on Web)

On Android:
- fixes a (backend) crash when client sends candidate before sending sdp
offer by queueing the candidates - on iOS it's not needed as sdp offer
is created on the same thread as candidates. In my opinion it's more of
a backend issue but it's quicker to workaround it on our side.
- fixes a bug when using `changeWebRTCLoggingSeverity` (found out when
debugging something)
- changes number strategy when serializing to json, so that it sends
integers instead of floats

On iOS:
- removed some dead code with `queuedRemoteCandidates`

## Motivation and Context


https://linear.app/swmansion/issue/FCE-555/investigate-work-required-to-support-elixirwebrtc-in-mobile-app

## How has this been tested?

- Get a backend with ex-webrtc (built locally from
`ex_webrtc_experimental` of fishjam or deployed to fishjam.work)
- Open fishjam dashboard
- Test only on physical devices, emulators won't work - currently
backend has fixed h264 encoding which is unsupported on android emulator
- Test different combinations: two mobiles, mobile + web
- Test connecting in different order, screencast, disconnecting and
reconnecting, adding and removing tracks

---------

Co-authored-by: Miłosz Filimowski <MiloszFilimowski@users.noreply.github.com>
Co-authored-by: Milosz Filimowski <milosz.filimowski@swmansion.com>