[FCE-806] Introduce connection hook (#220)
## Description

This PR introduces new hook `useFishjamConnection`. It aims to replace
useStatus*** hooks, so only one hook can be used in real apps. It also
share `join/leaveRoom` methods.

## Motivation and Context

After some discussions with external developers, it turned out that
having static methods in our API is a little confusing in react world.
So I decided to include them in hook, for better discoverability. Same
hook will also share connections status. So for most cases, this will be
only hook that has to be used to connect with Fishjam.

## How has this been tested?

This scenario was tested in example app. I does not change logic itself
(or maybe a little), so most important part is how new API is
structured.

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)

## Checklist:

- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.

## Screenshots (if appropriate)

---------

Co-authored-by: Maksymilian Galas <maksymilian.galas@swmansion.com>