[FCE-1126] Update publish docs action (#308)
## Description
Update all actions related to publishing docs for this project

## Motivation and Context
As of today,
[actions/upload-artifact](https://github.com/actions/upload-artifact) is
no longer available. So we have to update all actions that depend on it.


https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/

## How has this been tested?

It will be tested by merging and checking if action will run correctly
(it is not 'mission critical', so it can fail)

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)

## Checklist:

- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.

## Screenshots (if appropriate)