Feature/video player example (#408)
## Description

Added a new example demonstrating the use of a video player component.
This example showcases how to integrate and control video playback
within the Fishjam.

## Motivation and Context

This change is required to provide a practical reference for
implementing video playback using the video player component. It helps
developers understand how to integrate and control video playback within
the app. This addition improves the documentation and learning
experience, especially for those working with media features.

There is no related open issue, this is an enhancement for demonstration
purposes.

## How has this been tested?

Real Devices - iPhone 15 Pro and Samsug S23

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.