fix: podlife generation (#429)
## Description

Right now our script is required to always be called with --clean.

## Motivation and Context

On iOS, lines starting with # (Podfile comments) are skipped.
On Android, lines starting with # or / are ignored — the latter covers
comment styles used in Gradle files.
So overall, the Android version is a bit more flexible since it accounts
for different types of comments.

## How has this been tested?

Generated few times prebuild

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.