feat: add option to configure display name of ios screen broadcast extension (#527)
## Description

This PR adds the option to configure the display name of the extension.
This display name is shown in the native pop-up after starting to share
the screen.

## Motivation and Context

Currently, our app shows `FishjamScreenBroadcast` in the native popup,
which can be confusing for our users, as they have no idea what Fishjam
is.

## Documentation impact

I am not sure about this. I didn't find any related documentation to
update. Similar to tests, I have tested it in the
`internal/fishjam-chat` by running prebuild and checking the
`Info.plist` file manually.

- [ ] Documentation update required
- [ ] Documentation updated [in another PR](_)
- [ ] No documentation update required

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)